.dagger/pyproject.toml
[project]
name = "monorepo-dagger"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dagger-io",
    "tomli>=2.2.1"
]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
